moved string formatting methods to static class
